third_party/WebKit/LayoutTests/imported/wpt/html/semantics/document-metadata/the-base-element/base_href_specified-expected.txt

This is a testharness.js-based test.
FAIL The href attribute of the base element is specified assert_equals: The href attribute of the base element is incorrect. expected "http://www1.{{host}}:{{ports[http][0]}}/" but got ""
FAIL The src attribute of the img element must relative to the href attribute of the base element assert_equals: The src attribute of the img element is incorrect. expected "http://www1.{{host}}:{{ports[http][0]}}/test.ico" but got ""
Harness: the test ran to completion.
